dom/base/test/test_bug907892.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=907892
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 907892</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 907892 **/
  SimpleTest.waitForExplicitFinish();

  var expectedMessages = 2;
  window.onmessage = function (ev) {
    if (ev.data.sandboxed) {
      ok(ev.data.threw,
         "Should have thrown when setting document.domain in sandboxed iframe");
    } else {
      ok(!ev.data.threw,
         "Should not have thrown when setting document.domain in iframe");
    }

    --expectedMessages;
    if (expectedMessages == 0) {
      SimpleTest.finish();
    }
  };
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=907892">Mozilla Bug 907892</a>
<p id="display"></p>
<div id="content" style="display: none">
<!-- Set all the sandbox flags to "allow" to make sure we cover that case -->
<iframe
   sandbox="allow-same-origin allow-scripts allow-forms allow-top-navigation alllow-pointer-lock"
   src="http://test1.example.org/tests/dom/base/test/file_bug907892.html?sandboxed">
</iframe>
<iframe
   src="http://test1.example.org/tests/dom/base/test/file_bug907892.html?normal">
</iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
